% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_disclosure_control.R
\name{apply_disclosure_control}
\alias{apply_disclosure_control}
\title{Apply disclosure control to standard table}
\usage{
apply_disclosure_control(df)
}
\arguments{
\item{df}{Dataframe in the standard uptake data table structure.}
}
\value{
Dataframe
}
\description{
\code{apply_disclosure_control()} applies disclosure control to data in the standard data structure.
For the denominator and numerator columns, it applies the \code{replace_and_round()} function. It then
recalculates the indicator column using the rounded values.
}
