% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_leaflet_map}
\alias{create_leaflet_map}
\alias{create_map}
\title{Create a heat map in leaflet.}
\usage{
create_leaflet_map(df, input_geog_tabs)

create_map(df, unit)
}
\arguments{
\item{df}{Dataframe}

\item{input_geog_tabs}{Area level (e.g. ICB, provider) to plot.}

\item{unit}{Area level (e.g. ICB, provider) to plot.}
}
\value{
Leaflet object
}
\description{
\code{create_leaflet_map()} runs \code{create_map()} at the right geographical level (ICB or sub-ICB).

\code{create_map()} contains the code to create a map in leaflet.

Currently only compatible with proportions.
}
