% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_functions.R
\name{get_stat_title_text}
\alias{get_stat_title_text}
\title{Create caption for uptake graphs}
\usage{
get_stat_title_text(
  df,
  input_explore_tabs,
  selected_area_level,
  selected_area_name = NULL,
  geog_bar = FALSE
)
}
\arguments{
\item{df}{Dataframe}

\item{input_explore_tabs}{What data to plot (e.g. All, ICB, Ethnicity).}

\item{selected_area_level}{England or region/ICB/provider}

\item{selected_area_name}{Defaults to NULL. If a specific area is selected, e.g. Greater Manchester ICB,
plots that area's data only.}

\item{geog_bar}{Defaults to FALSE. If true, dates only cover last year.}
}
\value{
String
}
\description{
Creates a caption for uptake graphs
}
