% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_disclosure_control.R
\name{replace_and_round}
\alias{replace_and_round}
\title{Suppress small numbers and round to 5}
\usage{
replace_and_round(x)
}
\arguments{
\item{x}{Number to round or suppress}
}
\value{
Numeric
}
\description{
\code{replace_and_round()} replaces all values between 1 and 10 (inclusive) with 10
and rounds all numbers above 10 to the nearest 5.
}
\examples{
replace_and_round(6)

}
