% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptakegraphs-package.R
\docType{package}
\name{uptakegraphs-package}
\alias{uptakegraphs}
\alias{uptakegraphs-package}
\title{uptakegraphs: Functions used by the Impact team}
\description{
Plotting and cleaning functions for the Impact team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vtwkam1/uptakegraphs}
  \item Report bugs at \url{https://github.com/vtwkam1/uptakegraphs/issues}
}

}
\author{
\strong{Maintainer}: Vanessa Kam \email{vanessa.kam@nice.org.uk}

}
\keyword{internal}
