% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_leaflet_map.R
\name{create_leaflet_map}
\alias{create_leaflet_map}
\title{Create leaflet map using create_map()}
\usage{
create_leaflet_map(metric_df, input_geog_tabs)
}
\description{
Create leaflet map using create_map()
}
